+++ test_unstructured/partition/test_msg.py
@@ -39,11 +39,12 @@ def test_partition_msg_from_filename():
     assert elements[0].metadata == ElementMetadata(
         filename=filename,
         date="2022-12-16T17:04:16-05:00",
-        page_number=None,
+        page_number=1,
         url=None,
         sent_from=["Matthew Robinson <mrobinson@unstructured.io>"],
         sent_to=["Matthew Robinson (None)"],
         subject="Test Email",
+        filetype="application/vnd.ms-outlook",
     )
 
 
@@ -57,6 +58,8 @@ def test_partition_msg_from_filename_with_text_content(monkeypatch):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.msg")
     elements = partition_msg(filename=filename)
     assert str(elements[0]) == "Here is an email with plain text."
+    assert elements[0].metadata.filename == "fake-email.msg"
+    assert elements[0].metadata.file_directory == EXAMPLE_DOCS_DIRECTORY
 
 
 def test_partition_msg_raises_with_missing_file():