+++ test_unstructured/partition/test_pdf.py
@@ -118,7 +118,6 @@ def test_partition_pdf_local(monkeypatch, filename, file):
     )
 
     partition_pdf_response = pdf._partition_pdf_or_image_local(filename, file)
-    assert partition_pdf_response[0].type == "Title"
     assert partition_pdf_response[0].text == "Charlie Brown and the Great Pumpkin"
 
 
@@ -168,7 +167,7 @@ def test_partition_pdf(url, api_called, local_called, monkeypatch):
         attribute="_partition_via_api",
         new=mock.MagicMock(),
     ), mock.patch.object(pdf, "_partition_pdf_or_image_local", mock.MagicMock()):
-        pdf.partition_pdf(filename="fake.pdf", url=url)
+        pdf.partition_pdf(filename="fake.pdf", strategy="hi_res", url=url)
         assert pdf._partition_via_api.called == api_called
         assert pdf._partition_pdf_or_image_local.called == local_called
 
@@ -202,9 +201,18 @@ def test_partition_pdf_with_template(url, api_called, local_called, monkeypatch)
         attribute="_partition_via_api",
         new=mock.MagicMock(),
     ), mock.patch.object(pdf, "_partition_pdf_or_image_local", mock.MagicMock()):
-        pdf.partition_pdf(filename="fake.pdf", url=url, template="checkbox")
+        pdf.partition_pdf(filename="fake.pdf", strategy="hi_res", url=url, template="checkbox")
         assert pdf._partition_via_api.called == api_called
         assert pdf._partition_pdf_or_image_local.called == local_called
+
+
+def test_partition_pdf_with_auto_strategy(filename="example-docs/layout-parser-paper-fast.pdf"):
+    elements = pdf.partition_pdf(filename=filename, strategy="auto")
+    titles = [el for el in elements if el.category == "Title" and len(el.text.split(" ")) > 10]
+    title = "LayoutParser: A Uniﬁed Toolkit for Deep Learning Based Document Image Analysis"
+    assert titles[0].text == title
+    assert titles[0].metadata.filename == "layout-parser-paper-fast.pdf"
+    assert titles[0].metadata.file_directory == "example-docs"
 
 
 def test_partition_pdf_with_page_breaks(filename="example-docs/layout-parser-paper-fast.pdf"):