+++ test_unstructured_ingest/test-ingest-biomed-api.sh
@@ -25,11 +25,12 @@ if [[ "$(find test_unstructured_ingest/expected-structured-output/biomed-ingest-
 fi
 
 PYTHONPATH=. ./unstructured/ingest/main.py \
-   --metadata-exclude filename \
+   --metadata-exclude filename,file_directory \
    --biomed-api-from "2019-01-02" \
    --biomed-api-until "2019-01-02+00:03:10" \
    --structured-output-dir biomed-ingest-output-api  \
    --num-processes 2 \
+   --partition-strategy hi_res \
    --reprocess \
    --verbose \
    --re-download \