+++ test_unstructured_ingest/test-ingest.sh
@@ -10,6 +10,7 @@ export OMP_THREAD_LIMIT=1
 
 ./test_unstructured_ingest/test-ingest-s3.sh
 ./test_unstructured_ingest/test-ingest-azure.sh
+./test_unstructured_ingest/test-ingest-discord.sh
 ./test_unstructured_ingest/test-ingest-github.sh
 ./test_unstructured_ingest/test-ingest-gitlab.sh
 ./test_unstructured_ingest/test-ingest-wikipedia.sh