+++ unstructured/documents/elements.py
@@ -1,4 +1,8 @@
+from __future__ import annotations
+
+import datetime
 import hashlib
+import os
 import pathlib
 from abc import ABC
 from dataclasses import dataclass
@@ -14,7 +18,9 @@ class NoID(ABC):
 @dataclass
 class ElementMetadata:
     filename: Optional[str] = None
+    file_directory: Optional[str] = None
     date: Optional[str] = None
+    filetype: Optional[str] = None
 
     # Page numbers currenlty supported for PDF, HTML and PPT documents
     page_number: Optional[int] = None
@@ -34,12 +40,30 @@ class ElementMetadata:
         if isinstance(self.filename, pathlib.Path):
             self.filename = str(self.filename)
 
+        if self.filename is not None:
+            file_directory, filename = os.path.split(self.filename)
+            self.file_directory = file_directory or None
+            self.filename = filename
+
     def to_dict(self):
         return {key: value for key, value in self.__dict__.items() if value is not None}
 
     @classmethod
     def from_dict(cls, input_dict):
         return cls(**input_dict)
+
+    def merge(self, other: ElementMetadata):
+        for k in self.__dict__:
+            if getattr(self, k) is None:
+                setattr(self, k, getattr(other, k))
+        return self
+
+    def get_date(self) -> Optional[datetime.datetime]:
+        """Converts the date field to a datetime object."""
+        dt = None
+        if self.date is not None:
+            dt = datetime.datetime.fromisoformat(self.date)
+        return dt
 
 
 class Element(ABC):