+++ unstructured/ingest/interfaces.py
@@ -107,10 +107,10 @@ class BaseIngestDoc(ABC):
         """Write the structured json result for this doc. result must be json serializable."""
         pass
 
-    def partition_file(self):
+    def partition_file(self, **partition_kwargs):
         if not self.standard_config.partition_by_api:
             logger.debug("Using local partition")
-            elements = partition(filename=str(self.filename))
+            elements = partition(filename=str(self.filename), **partition_kwargs)
             return convert_to_dict(elements)
 
         else:
@@ -129,12 +129,12 @@ class BaseIngestDoc(ABC):
 
             return response.json()
 
-    def process_file(self):
+    def process_file(self, **partition_kwargs):
         if self.standard_config.download_only:
             return
         logger.info(f"Processing {self.filename}")
 
-        isd_elems = self.partition_file()
+        isd_elems = self.partition_file(**partition_kwargs)
 
         self.isd_elems_no_filename = []
         for elem in isd_elems: