+++ unstructured/partition/docx.py
@@ -5,6 +5,10 @@ from typing import IO, BinaryIO, List, Optional, Union, cast
 
 import docx
 import pypandoc
+from docx.oxml.shared import qn
+from docx.text.paragraph import Paragraph
+from docx.text.run import Run
+from tabulate import tabulate
 
 from unstructured.cleaners.core import clean_bullets
 from unstructured.documents.elements import (
@@ -13,9 +17,11 @@ from unstructured.documents.elements import (
     ElementMetadata,
     ListItem,
     NarrativeText,
+    Table,
     Text,
     Title,
 )
+from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
 from unstructured.partition.common import exactly_one, spooled_to_bytes_io_if_needed
 from unstructured.partition.text_type import (
     is_bulleted_text,
@@ -61,6 +67,36 @@ STYLE_TO_ELEMENT_MAPPING = {
 }
 
 
+def _get_paragraph_runs(paragraph):
+    """
+    Get hyperlink text from a paragraph object.
+    Without this, the default runs function skips over hyperlinks.
+
+    Args:
+        paragraph (Paragraph): A Paragraph object.
+
+    Returns:
+        list: A list of Run objects.
+    """
+
+    # Recursively get runs.
+    def _get_runs(node, parent):
+        for child in node:
+            # If the child is a run, yield a Run object
+            if child.tag == qn("w:r"):
+                yield Run(child, parent)
+            # If the child is a hyperlink, search for runs within it recursively
+            if child.tag == qn("w:hyperlink"):
+                yield from _get_runs(child, parent)
+
+    return list(_get_runs(paragraph._element, paragraph))
+
+
+# Add the runs property to the Paragraph class
+Paragraph.runs = property(lambda self: _get_paragraph_runs(self))
+
+
+@add_metadata_with_filetype(FileType.DOCX)
 def partition_docx(
     filename: Optional[str] = None,
     file: Optional[Union[IO, SpooledTemporaryFile]] = None,
@@ -92,13 +128,45 @@ def partition_docx(
 
     metadata_filename = metadata_filename or filename
     elements: List[Element] = []
-    for paragraph in document.paragraphs:
-        element = _paragraph_to_element(paragraph)
-        if element is not None:
-            element.metadata = ElementMetadata(filename=metadata_filename)
-            elements.append(element)
+    table_index = 0
+
+    for element_item in document.element.body:
+        if element_item.tag.endswith("tbl"):
+            table = document.tables[table_index]
+            html_table = _convert_table_to_text(table, as_html=True)
+            text_table = _convert_table_to_text(table, as_html=False)
+            element = Table(text_table)
+            if element is not None:
+                element.metadata = ElementMetadata(filename=metadata_filename)
+                element.metadata = ElementMetadata(text_as_html=html_table)
+                elements.append(element)
+            table_index += 1
+        elif element_item.tag.endswith("p"):
+            paragraph = docx.text.paragraph.Paragraph(element_item, document)
+            para_element: Optional[Text] = _paragraph_to_element(paragraph)
+            if para_element is not None:
+                para_element.metadata = ElementMetadata(filename=metadata_filename)
+                elements.append(para_element)
 
     return elements
+
+
+def _convert_table_to_text(table, as_html):
+    """
+    Convert a table object from a Word document to an HTML table string using the tabulate library.
+
+    Args:
+        table (Table): A Table object.
+        as_html (bool): Whether to return the table as an HTML string (True) or a
+            plain text string (False)
+
+    Returns:
+        str: An table string representation of the input table.
+    """
+    fmt = "html" if as_html else "plain"
+    headers = [cell.text for cell in table.rows[0].cells]
+    data = [[cell.text for cell in row.cells] for row in table.rows[1:]]
+    return tabulate(data, headers=headers, tablefmt=fmt)
 
 
 def _paragraph_to_element(paragraph: docx.text.paragraph.Paragraph) -> Optional[Text]: