+++ unstructured/partition/msg.py
@@ -4,12 +4,14 @@ from typing import IO, Dict, List, Optional
 import msg_parser
 
 from unstructured.documents.elements import Element, ElementMetadata
+from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
 from unstructured.partition.common import exactly_one
 from unstructured.partition.email import convert_to_iso_8601
 from unstructured.partition.html import partition_html
 from unstructured.partition.text import partition_text
 
 
+@add_metadata_with_filetype(FileType.MSG)
 def partition_msg(
     filename: Optional[str] = None,
     file: Optional[IO] = None,
@@ -39,15 +41,14 @@ def partition_msg(
     else:
         elements = partition_text(text=text)
 
-    metadata = build_msg_metadata(msg_obj)
-    metadata.filename = filename
+    metadata = build_msg_metadata(msg_obj, filename)
     for element in elements:
         element.metadata = metadata
 
     return elements
 
 
-def build_msg_metadata(msg_obj: msg_parser.MsOxMessage) -> ElementMetadata:
+def build_msg_metadata(msg_obj: msg_parser.MsOxMessage, filename: Optional[str]) -> ElementMetadata:
     """Creates an ElementMetadata object from the header information in the emai."""
     email_date = getattr(msg_obj, "sent_date", None)
     if email_date is not None:
@@ -66,6 +67,7 @@ def build_msg_metadata(msg_obj: msg_parser.MsOxMessage) -> ElementMetadata:
         sent_from=sent_from,
         subject=getattr(msg_obj, "subject", None),
         date=email_date,
+        filename=filename,
     )